checked & update live URL in these methode

+++ src/Pages/MyList/List/List.jsx
@@ -2,18 +2,45 @@ import { useContext } from "react";
 import { MdEditSquare } from "react-icons/md";
 import { Link } from "react-router-dom";
 import { AuthContext } from "../../../Provider/AuthProvider";
+import Swal from "sweetalert2";
+import PropTypes from 'prop-types'
+
 const List = ({ spot }) => {
     const {listedSpots} = useContext(AuthContext);
     console.log(listedSpots);
    
     const handleSpotsDelete = () => {
-        fetch(`https://world-tour-server-ten.vercel.app/spots/${spot._id}` , {
-            method: 'DELETE'
-        })
-        .then( ()=> {
-            console.log('successfully deleted');
-            window.location.reload();
-        })
+
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+          
+              fetch(`https://world-tour-server-ten.vercel.app/spots/${spot._id}` , {
+                  method: 'DELETE'
+              })
+              .then(res => res.json())
+              .then((data) => {
+                console.log(data);
+                if (data.deletedCount > 0) {
+                  Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success",
+                  });
+                  setTimeout(()=>{
+                  window.location.reload();
+                  }, 1000)
+                }
+              })
+          }
+        });
     }
     return (
         <>
@@ -55,3 +82,6 @@ const List = ({ spot }) => {
 };
 
 export default List;
+List.propTypes = {
+    spot: PropTypes.object
+}
